ml_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
python-multipart==0.0.9
joblib==1.5.3
scikit-learn==1.2.2
huggingface_hub==0.36.2
scipy==1.17.1
numpy==1.26.4
